Rename getRegistry and Published var
