+++ scripts/publish/src/getRegistryVersion.js
